backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for WOC Prototype"
authors = ["JuanseHevia <jh216@rice.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-restful = "^0.3.10"
flask-jwt-extended = "^4.6.0"
psycopg2-binary = "^2.9.9"
werkzeug = "^3.0.3"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
flask-cors = "^4.0.1"
six = "^1.16.0"
sqlalchemy = "^2.0.30"
pytest = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
